drop shape locator
