+++ packages/discovery/src/discovery/runDiscovery.ts
@@ -9,7 +9,6 @@ import { printSharedModuleInfo } from '../utils/printSharedModuleInfo'
 import { DiscoveryLogger } from './DiscoveryLogger'
 import { OverwriteCacheWrapper } from './OverwriteCacheWrapper'
 import type { Analysis } from './analysis/AddressAnalyzer'
-import { ShapeLocator } from './analysis/ShapeLocator'
 import { TEMPLATES_PATH, TemplateService } from './analysis/TemplateService'
 import type { ConfigReader } from './config/ConfigReader'
 import type { DiscoveryConfig } from './config/DiscoveryConfig'
@@ -79,13 +78,12 @@ export async function runDiscovery(
   }
 
   const templateService = new TemplateService(paths.discovery)
-  const shapeLocator = new ShapeLocator(templateService)
 
   printTemplatization(
     logger,
     result,
     !!config.verboseTemplatization,
-    shapeLocator,
+    templateService,
   )
 }
